+++ reghelp_client/__init__.py
@@ -33,7 +33,7 @@ from .models import (
     VoipStatusResponse,
 )
 
-__version__ = "1.3.4"
+__version__ = "1.4.0"
 __all__ = [
     "RegHelpClient",
     "BalanceResponse",